Release v1.5.0 -> Release v1.6.0 (#585)
* CAS-488: Voting Selection Style Updates (#568)

* update button

* create component

* export component

* extract component

* add icons

* create components

* extract components

* create const

* update icon

* add fade

* export component

* create component

* update style

* create component

* use CommunityName component

* add collapse button

* use StyledStatusPill instead of StatusLabel

* change borders

* fix merge

* fix merge

* update class names for modal

* move component

* create Modal component

* updates

* updates

* update casting vote modal

* updates

* updates

* remove check

* update

* fix folder name (#577)

* rename folder

* rename folder
