+++ frontend/packages/client/src/components/Proposal/VoteOptions/index.js
@@ -1,0 +1,1 @@
+export { default as VoteOptions } from './VoteOptions';
