+++ frontend/packages/client/src/components/Proposal/index.js
@@ -1,4 +1,4 @@
 export { default as CancelProposalModalConfirmation } from './CancelProposalModalConfirmation';
 export { default as ProposalStatus } from './ProposalStatus';
-export { default as VoteOptions } from './VoteOptions';
+export { VoteOptions } from './VoteOptions';
 export { default as HeaderNavigation } from './HeaderNavigation';
