+++ frontend/packages/client/src/components/modals/index.js
@@ -1,0 +1,6 @@
+export { default as Modal } from './Modal';
+export { default as Error } from './Error';
+export { default as VoteConfirmation } from './VoteConfirmation';
+export { default as CastingVote } from './CastingVote';
+export { default as VoteConfirmed } from './VoteConfirmed';
+export { default as CancelProposal } from './CancelProposal';
